code refactoring

+++ src/main/java/com/article/service/repository/ArticleRepo.java
@@ -3,5 +3,5 @@ package com.article.service.repository;
 import com.article.service.entity.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleRepo extends JpaRepository<Article, Integer> {
+public interface ArticleRepo extends JpaRepository<Article, Long> {
 }
